Using Inventory and extracting logic from Handler
